Finish Computational Project
Finished project, including polishing existing code, adding full documentation, attaching final written report, and creating additional routines to determine the number density of particles for a given temperature and chemical potential. This information is used to calculate the theoretical energy density of the Fermi gas and, following Stafford+ 1999, the surface energy for comparison with the computed values.